ci: split SHASUM summary comment workflow into two
commenting on a PR requires permissions which don't work from forks of
public repos. Circumvent this by using a second workflow which runs via
`on_workflow`, where the correct permissions are obtained.

Previously, the workflow would run the shasum-summary tool and store the
comment in a file. The file contents would then be used as the comment
body. Commenting however fails due to external forks not having the
permissions to comment.

Now, the workflow is split into two parts. First, the comment (and PR
number) is stored in file which is uploaded as build-artifact. Upon
completion, a second workflow starts, downloads the build-artifact, and
comments on the PR.